requirements.txt
bitarray==2.7.3
blake3==0.3.3
cffi==1.15.1
cytoolz==0.12.1
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
grpcio==1.53.0
grpcio-tools==1.53.0
hexbytes==0.3.0
iniconfig==2.0.0
packaging==23.0
parsimonious==0.9.0
pluggy==1.0.0
protobuf==4.22.1
pycparser==2.21
pycryptodome==3.17
PyNaCl==1.5.0
pytest==7.3.0
pytest-mock==3.10.0
python-dotenv==1.0.0
regex==2023.3.23
rlp==3.0.0
toolz==0.12.0
